Create modbusTest.nut
As of August,2013 Modbus/RS485 wasn't supported by the ElectricImp library, though it is said to be in the works.
This is for use with Electric Imps WiFi device. Currently tested on an IMP001 - see pinout for RS485 connection.
The top part of this file is loaded into Electric Imps Device window, that runs on the WiFi IMP001 on site.
The bottom part is loaded into the Agent and runs on the internet.